chatbot-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.4.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.9